Refactor API call types and data shape
This preps for adding an alternate backend, that
uses the Supabase JS client instead of Axios calls.

The aim is to use shared types for API responses and errors
within the UI code, and push any API differences down lower.

For signup and login, I had to parse the API call
result object, and rethrow any errors to fix a bug
where users could attempt to sign up with a duplicate
username but think they were successful.