Ignore data conflicts for newly-created users
This is (tentatively) done to help development of
an alternate backend. The frontend seems to expect
a fully-formed data object for a new user, but
I'm trying to avoid the backend needing to know
how to create this whole thing.

This change aims to allow a new user to ignore data
conflicts, allow the frontend to create and update
the initial data reponses, and things flow from there.